- Allowed images to be uploaded to the markdown textarea    -> Added a boolean property to enable or disable it per-control    -> Passed all new properties through the FormTextarea    -> The file uploader component opens up in a modal, which also as a URL input    -> When a queued file is clicked, it will be added to the post as a data variable - Exported textarea props and used them in TextareaControlConfig instead of redefining them there - Moved 'disabled' from textarea htmlprops into controlled props - Added message overrides to Textarea    -> A disabled control with useMarkdown() disables the buttons - Made onUpload() and onRemove() properties of the FileUploader optional - Added ability for pending files in FileUploader to be clicked on (selected), and for those selections to be persisted - Added a Modal component    -> Exported it as a library component - Disallowed images from running outside the container in MarkdownRenderer

+++ src/common/form/controls/attachment.view.tsx
@@ -19,7 +19,7 @@ export function FormAttachmentControl<T>(
       defaultValue={props.defaultValue}
       supportedFileExtensions={props.supportedFileExtensions}
       onUpload={(files: File[]) => {
-        props.onUpload(files)
+        if (props.onUpload) props.onUpload(files)
         if (props.isMultiple) {
           props.saveValueToState(
             props.propertyPath,
@@ -35,7 +35,7 @@ export function FormAttachmentControl<T>(
         if (shouldValidate) props.validate(null)
       }}
       onRemove={(filename: string, index: number) => {
-        props.onRemove(filename, index)
+        if (props.onRemove) props.onRemove(filename, index)
         if (props.isMultiple) {
           props.saveValueToState(
             props.propertyPath,
